+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Changes
 =======
 
-0.9.14 (unreleased)
--------------------
+1.0 (unreleased)
+----------------
 
 - Remove usage of ``Nodespaces`` behavior.
   [rnix]